% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif.R
\name{get_hl_alpha}
\alias{get_hl_alpha}
\title{Get Highlight Alpha}
\usage{
get_hl_alpha(default = 0.4)
}
\arguments{
\item{default}{Default value if not set (0.4).}
}
\value{
A numeric value representing the alpha level.
}
\description{
Gets the current alpha value for motif highlighting.
}
\examples{
get_hl_alpha()
}
