% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style.R
\name{format_linkage_label}
\alias{format_linkage_label}
\title{Format Linkage Labels}
\usage{
format_linkage_label(labels)
}
\arguments{
\item{labels}{A character vector of linkage labels.}
}
\value{
A character vector with Greek letters.
}
\description{
Converts 'a'/'b' to Greek letters alpha/beta.
}
