% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif.R
\name{highlight_motif}
\alias{highlight_motif}
\title{Identify Motif in Glycan Graph}
\usage{
highlight_motif(graph, motif_str)
}
\arguments{
\item{graph}{The main glycan graph (igraph).}

\item{motif_str}{The motif string (IUPAC format).}
}
\value{
A graph with an added `alpha` vertex attribute (1 for motif, 0.4 for others).
}
\description{
Finds occurrences of a motif subsequence in the main glycan graph and marks them.
}
