% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style.R
\name{match_snfg_style}
\alias{match_snfg_style}
\title{Match SNFG Style to Graph Nodes}
\usage{
match_snfg_style(graph)
}
\arguments{
\item{graph}{A `tbl_graph` or `igraph` object.}
}
\value{
A graph object with additional columns: `snfg_shape` (numeric starshape) and `snfg_fill` (color).
}
\description{
Maps glycan node labels to SNFG shapes and colors.
}
