% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_glycan}
\alias{read_glycan}
\title{Read Glycan Data}
\usage{
read_glycan(x, format = "iupac")
}
\arguments{
\item{x}{A character string representing the glycan structure.}

\item{format}{The format of the string. Currently only "iupac" is supported.}
}
\value{
A `tbl_graph` object representing the glycan structure.
}
\description{
Parses a glycan string (IUPAC condensed format) and returns a graph object.
}
