% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif.R
\name{set_hl_alpha}
\alias{set_hl_alpha}
\title{Set Highlight Alpha}
\usage{
set_hl_alpha(value = 0.4)
}
\arguments{
\item{value}{Numeric value for alpha (default 0.4).}
}
\description{
Sets the alpha value for motif highlighting.
}
